Renamed helpers to aptible
