+++ aptible/tokens.go
@@ -1,4 +1,4 @@
-package helpers
+package aptible
 
 import (
 	"encoding/json"
